requirements.txt
monai==1.1.0
scikit-learn
tqdm
medpy
tensorboard
six
setuptools==59.5.0
PyYAML
numpy==1.23.5
